Move message definitions functions to loader module and clean up definitions module
